requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.0
aiohttp>=3.9.0
numpy>=1.26.0
matplotlib>=3.8.0
pandas>=2.1.0
python-dotenv>=1.0.0
